isbnv2test removes non-numerical or X characters
seems to work but I worked on it so long it got sloppy.  If "123456789Z" is entered, it deletes Z and the test fails.  If "123456789X" is entered, X is turned into 10 and the test passes.  It also removes any non-numerical characters from the first 9 characters.  So if "123x56@89X" is entered, it will delete index[3] which is x, and will delete index[6] which is @, but will still translate X to 10.